Fixed bug where channel was not found if pygame.mixer was pre-intialized
If pygame.mixer.init() was run before Voice.speak() was called, then no channel was
assigned for the voice to play on. Now if pygame..mixer.init() was already called, then
an open channel is searched for and the voice is played on that. Otherwise a new
channel is created, and the voice is played on that.